[FTQC] Add offline byproduct correction support  (#7447)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**

[sc-91491]

**Background**

The `offline_byprod_correction` method allows us to move all `byproduct
ops` (`X`, `Z` or `XZ`) as well as `Pauli` operations (`I`, `X`, `Y`,
`Z`) of MBQC (Measurement Based Quantum Computing) to the end of a
`Clifford` circuit and the measurement result can be corrected per shot
based on the `Paulis` recorded in the function, i.e. the phase of
`Z-basis` measurement result would be flipped if there is a `X` recorded
at the target wire.

This function reduces the number of quantum gate operations for the MBQC
formalism by utilization of classical hardwares. This is achieved based
on the fact that a Cliiford gate ($C$) can conjugate a Pauli to a new
Pauli, i.e., $P_{new}C = CP_{old}$.

Note that `RZ` or `RotXZX` operators at the beginning of a circuit for
each wire is supported, given that no Paulis are commuted with those
operators. This would be important for the `StatePrep` operations.

**Implementations**

1. The `offline_byprod_correction` method accepts `2` arguments:  
- a `tape` contains only $qml.I$, $qml.X$, $qml.Y$, $qml.Z$, $qml.S$,
$qml.H$ and $qml.CNOT$;
- mid-circuit measurement results per shot.

2. The `offline_byprod_correction` method initializes three arrays: 
- `X` correction; 
- `Z` correction; 
- `phase` information for correction of each wire.

3. Then the `tape` and `mid_meas` mid-circuit measurement results are
parsed to build stack of `byprod` ops and the tape ops.

4. Then the commutate rules are applied to the byprod and the Paulis or
Clifford gates in the circuit.

5. Finally, the phase factor used to correct measurement results is
returned based on the recorded Paulis and measurement obs of the target
wire.

Note that a few integration tests, which are subject to be updated once
the transform pipeline is done, are added for circuits with Clifford and
Paulis.

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>
Co-authored-by: lillian542 <38584660+lillian542@users.noreply.github.com>
Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>