[Decomposition] The decompose transform accepts stopping_condition (#7531)
**Context:**

PennyLane currently relies on decompositions derived from a given
stopping_condition (or stopping_condition_shots) function in multiple
places within the broader code base. We want to be sure that the new
decompositions system can be used in every internal facet of PennyLane.

The `gate_set` and `stopping_condition` are two different concepts. The
`stopping_condition` is an acceptance criteria for operator instances
that need no further decomposition, but on top of this, the
decomposition graph requires a set of gate types as the target gate set.
With graph enabled, the `decompose` transform would take both `gate_set`
that is a set of gate types, and `stopping_condition` that takes
operator instances.

**Description of the Change:**

Adds a `stopping_condition` argument to the `decompose` transform.

- Graph Enabled
- `gate_set` must be an iterable of gate types and/or names or a
dictionary (weighted gate set)
- `stopping_condition` must be a callable that takes operator instances
and returns `True` or `False`
- If not provided, the default stopping condition is "whether the
operator is part of the gate set"
- Graph Disabled
- Either provide `gate_set` or `stopping_condition`, but not both (the
old system does not make the distinction, so we require that only one of
them is given to prevent ambiguity.
- If `gate_set` is provided, it is ultimately translated to a callable
stopping condition (same behaviour as before)
  - If `stopping_condition` is provided, it is used directly.

**Benefits:**

An instance-based stopping condition is supported in the new graph-based
decomposition system.

**Related GitHub Issues:**

[sc-92234]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>