No more hard restriction on `StatePrep` normalization (#7615)
**Context:**
As requested by https://github.com/PennyLaneAI/pennylane/issues/7259 we
would like to lift the hard restriction on `StatePrep` in the past. Now
users should be able to input a not normalized state with no Error atl
but a `UserWarning`

**Description of the Change:**

**Benefits:**
Fix a bug(?) that users do not expected to be blocked restrictively in
default setup.

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-88797]