[Decomposition] Add `work_wire_type` to all controlled operators. (#7612)
**Context:**

Now that more decomposition rules that use work wires are implemented,
more fine grained control of the work wire type is required. For
example, the decomposition of `PCPhase` passes some of its operation
wires to the `Controlled(PhaseShift)` as work wires. Now there is a
decomposition rule for a `Controlled(PhaseShift)` that uses work wires,
but it requires the work wire to be clean. This is not satisfied by the
wires passed from the `PCPhase`, therefore this decomposition rule is
not actually applicable.

We need to keep track of the work wire type in all controlled operations
now. Therefore, we need to add a `work_wire_type` argument to `ctrl`,
and a `work_wire_type` property to the general `Controlled` class.

In the long term, we aim to get rid of this entirely with dynamic qubit
allocation, but for the time being, this change is necessary to ensure
the correctness of decompositions.

**Description of the Change:**

- Adds `work_wire_type` to `ctrl`, `Controlled`,
`ControlledQubitUnitary`, and all decomposition rules of controlled
operators.
- Update the default work wire type to "dirty"

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-92556]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>