Fixed Qdrift test (#7621)
**Context:**
Integration tests for the Qdrift template were failing. These tests were
comparing the state produced after applying QDrift with the state
produced after applying its decomposition.

The source of the discrepancy was a bug in the tests themselves.
Specifically, the expected state was produced by multiplying the gates
in the expected decomposition of QDrift in the opposite order. The order
should be reversed from that of the decomposition.

This should have caused all of the tests to fail initially, however the
hamiltonians used while testing Qdrift had accidentally been chosen such
that most of their terms commuted with each other. Due to this, the
order of the decomposition didn't matter except for the few edge cases.

**Description of the Change:**
  - Updated the test hamiltonians so that they don't all commute. 
  - Reversed the order of the `expected_decomp` list

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Simone Gasperini <simone.gasperini4@unibo.it>