Elbow operator (#7472)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**

[sc-91810]

**Description of the Change:**

Adding the elbows operators described in:
https://arxiv.org/abs/1805.03662

![Screenshot 2025-05-23 at 10 19
00 AM](https://github.com/user-attachments/assets/886af59c-bb93-4fa2-a975-5471a223dc44)


**Benefits:**

More efficient decomposition in templates

**Possible Drawbacks:**

Right now we can't apply MCMs, we will have to update the operator in
the future for the Adjoint(elbow)
**Related GitHub Issues:**

---------

Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>
Co-authored-by: Utkarsh <utkarshazad98@gmail.com>
Co-authored-by: ANTH0NY <39093564+AntonNI8@users.noreply.github.com>
Co-authored-by: Soran Jahangiri <40344468+soranjh@users.noreply.github.com>