Caching defaults to auto with execute (#7644)
**Context:**

Caching can get expensive when we don't actually have duplicate tapes,
so it's not worth it when `max_diff=1` the vast majority of the time.
Also, it can cause confusion when it eliminates the expected randomness.

**Description of the Change:**

Defaults `cache="auto"` in `qml.execute` matching the behavior of
`QNode`.

*Always* warn about caching with finite shots.

**Benefits:**

Increased performance and reduced confusion.

**Possible Drawbacks:**

**Related GitHub Issues:**

Fixes #7643  [sc-92964]