Restrict `labs` dependencies to the functions that introduce them (#7650)
**Context:**
The labs module contains experimental features, some of which require
additional packages, causing dependencies that PennyLane itself does not
(want to) have.
Currently, these additional packages are imported as soon as any `labs`
functionality is imported.

**Description of the Change:**
Change the imports so that they only cause a problem if a function is
called that actually requires one of the additional dependencies.

**Benefits:**
Can use `labs.submodule_A` without installing dependencies of
`labs.submodule_B`.
Can use `labs.submodule_A.fun_1` without installing dependencies of
`labs.submodule_A.fun_2`.

**Possible Drawbacks:**
More cumbersome import structure. Not really a problem. We do this with
JAX a lot in PennyLane already.

**Related GitHub Issues:**
This is an alternative approach to #7649.

---------

Co-authored-by: Korbinian Kottmann <43949391+Qottmann@users.noreply.github.com>
Co-authored-by: Diego <67476785+DSGuala@users.noreply.github.com>