Fix `qml.FlipSign` operation (#7647)
Fix #7350.

The original example in the linked issue should actually raise an error
because `qml.FlipSign(2, wires=2)` is trying to flip the sign of the
basis state `2` (or `[1,0]` in binary repr) targeting a single wire.

Taking a deeper look into this, I also noticed that `qml.FlipSign`
supports targeting a sub system of the circuit as well. For instance,
with `n=[0]` and `wires=[0]` in a circuit with 2 wires, it flips the
sign of all the basis states where the first (left-most) qubit is 0
(i.e. 00 and 01).
```python
import pennylane as qml

num_wires = 2
dev = qml.device("default.qubit", wires=num_wires)

@qml.qnode(dev)
def circuit():
    for wire in range(num_wires):
        qml.Hadamard(wires=wire)
    qml.FlipSign(n=[0], wires=[0])
    return qml.state()
```
```
>>> circuit()
array([-0.5+0.j, -0.5+0.j,  0.5+0.j,  0.5+0.j])
```
However, this is currently not documented nor tested at all. If that is
the expected behaviour, I think it would make sense to include docs and
tests as part of this PR. Any feedback is welcome!

[sc-90122]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>