[Python Compiler] Restructuring `python_compiler` submodule (#7645)
* Update implementation of `xdsl_transform` to be slightly cleaner. Also
added `module_pass` and `name` attributes to the class that
`xdsl_transform` returns for easy access to the `ModulePass`.
* Restructure the whole `python_compiler` submodule (to see the
difference, I would recommend actually viewing the files in this PR's
branch rather than looking at the diffs. General ideas used:
* Moved function/class definitions out of `__init__.py` files and into
dedicated files.
  * Subdirectories with a single file were removed.
* Added `xdsl-` prefix to the pass names so that there is stricter
differentiation between catalyst and xdsl passes. This was motivated by
the fact that `merge-rotations` is the name for the the xdsl and
catalyst passes.

Happy to get feedback on the new structure and the new implementation of
`xdsl_transform`.

[sc-92828] [sc-92677]

---------

Co-authored-by: erick-xanadu <110487834+erick-xanadu@users.noreply.github.com>