Fix `qml.draw_mpl` with `work_wires` present (#7668)
**Context:**
`qml.draw_mpl` maps wires before drawing. 
It does not consider work wires, so that there can arise collisions
between mapped algorithmic wires and non-mapped work wires. This is
similar to the problem in #7580 (fixed in #7581).

**Description of the Change:**
Consider work wires in the wire mapping.
As the map is also used to decide which wires are drawn at all, we need
to now produce two maps: One for mapping (now incl. work wires) and one
for indicating which wires are drawn (excl. work wires).

**Benefits:**
fixed bug.

**Possible Drawbacks:**
n/a

**Related GitHub Issues:**
fixes #7659 .

[sc-93132]