Create `QNode._shots` to support `shots` updates (#7627)
**Context:**
A simple transform that takes one tape with certain `shots` to another
won't work, at least not gonna be replace previous `device.shots` as
expected, since the actual scope of shots information is way larger than
the scope where a transform takes effect. Specifically, if `set_shots`
wants to change a plain `shots` into a `shot_vector`, it won't work,
because the information about the shape of `shot_vector` should be
playing a role from the init of `QNode` and a transform cannot control
that far.

Therefore, we need to adjust the whole system a little bit. In the PR,
we are going to try make `shots` a secret property of `QNode`, but not
directly public to users. That is to say, one should be able to call
`QNode.update(shots=shots)` just as expected for the `set_shots`;
however, a user cannot do things like `QNode._shots` directly to access
or modify this part of information.

### !Note: Precedence of `QNode._shots`
 - If set_shots was applied, use shots from set shots
- If set_shots was applied and shots is in the call, we warn and use the
shots from set_shots
- If set_shots was not applied and shots is in the call, we use the
shots in the call


**Description of the Change:**
 - Let there be _shots along with _set_shots() inside QNode
- At the entry of QNode.construct(), override previous shots (from
device basically) with _shots
 - An update_shots(), and updated update() with shots=
- set_shots was moved from transforms/ to workflow/ beside qnode.py,
along with its tests extended and moved to test_qnode.py

**Benefits:**
- [ ] `shots` way more decoupled than before. Now `shots` almost has its
own decision to be set once via `set_shots` and all the following up
workflow just do whatever they were supposed to do as before
- [ ] `device.shots` more clearly overridden and could be easily
deprecated next quarter (?)

**Possible Drawbacks:**
`set_shots` does not exist as a `transform` any more. Well, it's hard to
say this is really a drawback though, but compared with current `master`
branch it is indeed a breaking.

**Related GitHub Issues:**
[sc-92752]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>