Dynamic qubit allocation version 2 (#7704)
**Context:**

This replaces PR #7663 .

**Description of the Change:**

Adds a allocation module with allocate, and deallocate functions for
working with dynamic qubits. The dynamic qubits themselves are special
class that is identified by a unique uuid. `allocate` returns a special
child of `Wires` that can act as a context manager with deallocation.

Allocate and Deallocate ops present the creation and freeing of these
wires in core pennylane.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-93180]

---------

Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>