Add call_graphs for PennyLane operators to support PennyLane to Qualtran wrapper functionality (#7536)
**Context:**
For the PL to Qualtran interoperability, we need to write some
`call_graphs()` for select operators that we have determined to be
useful & also have expensive decompositions. This allows us to use the
call graph functionality that exists in qualtran, which enables resource
estimation even for trillions of qubits. Important to note that relying
on just the PL decomposition would be similar to using specs/tracker,
which has performance issues. In this case, we are using a combination
of these pre-defined dictionaries and relying on the PL decomposition at
times, leading to resources that may not exactly match what you would
get from specs or tracker. This is fine though, because it is not
guaranteed that a bloq's decomposition and its call graph match exactly
-- this behaviour is not completely unexpected. However, we do make note
of it in the documentation for clarity.

**Description of the Change:**
Add call_graphs as well as associated tests for some operators. The call
graphs are either inspired by the original PL decomposition or the
resource decomposition, deferring to the latter when performance is an
issue.

**Benefits:**
Enables resource estimation for large circuits using Qualtran's UI

**Possible Drawbacks:**
May have inconsistent resource estimates when compared to
resources/tracker, due to the fact that not all operators have a call
graph defined.

**Related GitHub Issues:**

---------

Co-authored-by: Utkarsh <utkarshazad98@gmail.com>