Add `python_compiler` utilities for ease of use and testing (#7712)
* Add `inline_module` function that takes two modules as input, and
inlines the body of the first module into the second module.
* Add `inline_jit_to_module` function that takes a `jax.jit`-ed function
and xDSL module as inlines the xDSL module corresponding to the jitted
function into the user provided module. This function doesn't return
anything; it modifies the module in-place.
* Add `xdsl_from_qjit` function, which takes a QJIT-ed function as input
and returns a wrapper that, when called, returns an xDSL module.
[sc-93673] [sc-93672] [sc-93735]

---------

Co-authored-by: erick-xanadu <110487834+erick-xanadu@users.noreply.github.com>