Ensure parametric MCM objects are compatible with Jax-centric angle parameters (#7693)
**Context:** This PR fixes an issue where Jax-centric angle inputs were
unable with the the parametric MCM inputs. The issue was due to
hashability of array data, and the lack of unpacking to a fixed tuple.
To avoid failures, the hashing of the underlying data now relies on
conversion of the angle to a fixed array, which is then fed into an
instance of the
[hashlib](https://docs.python.org/3/library/hashlib.html) library's
implementations. The underlying numpy array follows the Python buffer
API, and so the data is directly readable for a hashed representation.

**Description of the Change:** As above.

**Benefits:** Ensures the angle parameter is compatible with Python
floats, numpy and Jax 0-dim float arrays, and even Jitted functions.

**Possible Drawbacks:** As this isn't anticipated to be used anywhere
outside of the MCM functionality, no concerns are currently raised for
collisions between hashes of different objects. However, due to the
inability to hash JITable objects, the implementation choice for relying
on the string angle may

**Related GitHub Issues:**