Fix trainability of classical preprocessing in transform programs (#7345)
**Context:**
`validate_device_wires` does not preserve the `trainable_params`
property of the tape if there are any measurements that do not have
`obs` or `wires`, like `qml.probs()`.

**Previous context:**
For differentiation-related transforms, the classical Jacobian needs to
be obtained during program transform execution.
For this, the classical preprocessing is traced throughout the transform
program, and the _trainable_ parameters of the processed tape, as
indicated by `trainable_params`, are returned from
`_classical_preprocessing`. This function can then be differentiated to
obtain the classical Jacobian.

Currently, the trainability of the tape created from a QNode is
maintained correctly. However, applying a transform program is not
guaranteed to set the correct `trainable_params` on the transformed
tape. This makes `_classical_preprocessing` return _all_ parameters,
instead of just those that should be trainable based on QNode-level
information. Consequently, the classical Jacobian has wrong dimensions.

**Description of the Change:**
Pass the `trainable_params` of the original tape to the new, modified
tape in `validate_device_wires`, if there is modification at all.

**In addition: ** Update the `trainable_params` of the tape transformed
via the transform program, so that `_classical_preprocessing` returns
trainable parameters only, even if some intermediate transform did not
maintain the information.

**Benefits:**
Bug fix for differentiation-related transforms with classical
cotransforms applied to QNodes.

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
Fixes #7356 
[sc-90203]