Move internal code of TransformProgram onto CotransformCache (#7750)
**Context:**

Currently `TransformProgram` has a lot of code for calculating the
classical jacobian and the argnums. All this code is rather tied to the
`CotransformCache` that holds onto the qnode, args, and kwargs.

By moving this code onto the `CotransformCache`, it simplifies
`TransformProgram` and opens up opportunities cleaning up
`TransformProgram.__call__`. But even if we don't clean up
`TransformProgram.__call__`, this still simplifies `TransformProgram`.

**Description of the Change:**

- Extracts all the code that interacts with `CotransformCache` *into*
`CotransformCache`.
- Cleans up imports in `transforms/core`

**Benefits:**

`TransformProgram` is more focused on it's real responsibilities.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: andrijapau <andrija.paurevic@xanadu.ai>
Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>