[Python Compiler] Do not use AnyAttr in AnyOf (#7772)
**Context:** Recent changes in xDSL restrict the use of AnyAttr in
AnyOf. See PR [here](https://github.com/xdslproject/xdsl/pull/4510)

**Description of the Change:** We don't need to say `AnyOf(AnyAttr(),
AnyAttr())`. We can just say `AnyAttr()` and it means the same thing. We
could make this check stricter by using the correct types which in most
of the cases are TensorTypes or MemrefTypes, but I think this is ok for
now.

**Benefits:** The latest version of xDSL works with the Python compiler.

**Possible Drawbacks:** None.

**Related GitHub Issues:**