Improvements to RS decomposition (#7770)
**Context:** Some improvements to the Ross-Selinger decomposition
method.

**Description of the Change:**
1. Added recipe for a tighter ellipse construction than before. This
allows for obtaining solutions with lesser `k` values. This also gives a
clear indicator (in terms of `ZeroDivisionError`) for the need for
higher precision. Support for that is not included here.
2. More efficient skew reduction. Modification to the step lemma to
increase its performance. A key change comes from the handling of
shifts, where larger shifts are now handled by $\sigma^2$, which turns
out to be a special grid operator. For the lower shifts, I changed the
formulation to handle shifts in the skew-reducing grid operators to
isolate the shift conjugation only to them. Also, add edge-case handling
with the `X/Z` operators.
3. Separate out the logic for testing the initial guesses, so that
`kmin` values don't affect the solution testing in their case.
4. Changed the name of the method to `gridsynth` from `rs` in the
`clifford_t_decomposition` transform.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:** [sc-94146, sc-93946]