Jax jit givens decomposition (#7823)
**Context:**
This PR is a combination of @erick-xanadu's previous
[PR1](https://github.com/PennyLaneAI/pennylane/pull/7013) and
[PR2](https://github.com/PennyLaneAI/pennylane/pull/7014).

**Description of the Change:**
It jax jits the `givens_matrix` from `BasisRotation` when it is within a
jit context, which significantly reduces the program size and
compilation time of workflows, e.g. the XAS benchmark.

**Benefits:**
This significantly reduces the time for XAS with compilation benchmark
([~5-6x](https://github.com/PennyLaneAI/pennylane-benchmarks/actions/runs/16158349592/job/45616094304#step:10:104)).
This undoes most of the performance degradation coming from the
bufferization updates in catalyst.

**Possible Drawbacks:**
Jax jitting is now part of the Pennylane internals

**Related GitHub Issues:**

[sc-94738]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>