Remove warning in the `qml.FlipSign` operation (#7875)
A warning is currently raised when creating an instance of the
`qml.FlipSign` operation passing an integer number as the `wires`
argument. This was introduced by [this
PR](https://github.com/PennyLaneAI/pennylane/pull/7647) to warn users
about the new `qml.FlipSign` behaviour fixing the previous wrong
interpretation of that argument.

The aim of this PR is to remove the warning.

[sc-93136]