New set shots (#7876)
**Context:**
`@partial(qml.set_shots, shots=...)` seems too complicated and
user-unfriendly. So we would like to have the alias that
`@qml.set_shots` got applied on the qnode directly.

**Description of the Change:**
Extended the fucntionality of the original `set_shots` to be able to
handle the case where it's used as a general decorator so that it
returned a fuction.

**Benefits:**
`@set_shots` is live now.
Plus, it co-exists with previous usage, so one can still do
`@partial(qml.set_shots, shots=...)` as wish.

**Possible Drawbacks:**
Maybe we can check if the recursive usage is dangerous? It smells not
bad to me but maybe there will be certain corner cases.

**Related GitHub Issues:**
[sc-95448]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>