Remove `shots` from kwargs for `cut_circuit_mc` (#7882)
**Context:**

`cut_circuit_mc` was doing some rather weird stuff with shots and the
qnode that is rather outdated and no longer necessary.

**Description of the Change:**

Remove `shots` from `cut_circuit_mc` and just take the shots from the
tape.

**Benefits:**

More straightforward, robust code.

**Possible Drawbacks:**

A full breaking change removal. But I'm not too concerned about it, and
it's a rather easy one to adapt to.

**Related GitHub Issues:**
[sc-95028]