Add `dialects` submodule to the Python compiler (#7897)
This PR adds a `dialects` submodule to `qml.compiler.python_compiler`.
Since we will have at least 3 dialects in the Python compiler soon (once
the Catalyst dialect is ported), it makes sense to start adding more
structure to the Python compiler.

Additionally, I updated the names of the dialect files to not use the
word "dialect":
* `mbqc_dialect.py` -> `mbqc.py`
* `quantum_dialect.py` -> `quantum.py`

As well as the same change to the names of the actual dialects:
* `QuantumDialect` -> `Quantum`
* `MBQCDialect` -> `MBQC`

Random change: I saw that there was a `pytest` related changelog entry
under the "Improvements" section of the changelog. This section is for
_user-facing_ improvements, so I moved it to the "Internal changes"
section.

[sc-95586]