Implement an AutoGraph-friendly `wraps` (#7889)
When using AutoGraph, generally source conversions are restricted to
"user-level" code, that is efforts are made to prevent the conversion of
library code. The reason is that library code often contains a lot of
Python logic that needs to execute during program capture, that cannot
and should not be converted into traceable code.

Using `functools.wraps` to wrap library functions around user-defined
functions circumvents this selective conversion mechanism, and can lead
to errors like the following:
https://github.com/PennyLaneAI/catalyst/issues/1824
The reason is that conversion decisions are made based on the module a
function came from. If a function is from a Catalyst or PennyLane
module, conversion does not take place. In scenarios like the one
described in the issue, the target function's `__module__` property is
overwritten so that the PennyLane wrapper function is converted by
AutoGraph, leading to an error.

In Catalyst, we implemented a solution by selectively updating
properties when wrapping functions (excluding the `__module__`
property), see [here for
example](https://github.com/PennyLaneAI/catalyst/blob/596d038c7d453a198218636635c7478849bfe8af/frontend/catalyst/api_extensions/callbacks.py#L348).
PennyLane lacks this fix which is introduced in this PR.

I focused mainly on finding uses of `functools.wraps` for "user-facing"
decorators (which would be affected by the issue described above), but
feel free to suggest others or corrections. I included functions from
`fourier`, `ftqc`, `op_math`, and `trotter` modules.

Partial fix for https://github.com/PennyLaneAI/catalyst/issues/1824.
[sc-93971]

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>