Fix `to_zx` transform to support Toffoli and CCZ gates (#7899)
At the moment, not all the gates available in PennyLane are properly
converted into their ZX-diagram representation by the `to_zx` transform.
This is the reason behind issues like #7669.

This PR implements a fix to extend the support to Toffoli and CCZ gates,
providing an explicit decomposition of these operations to the basis
gates set used for the conversion PL circuit --> PyZX graph.

[sc-93185]

---------

Co-authored-by: Marcus Edwards <94946848+comp-phys-marc@users.noreply.github.com>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>