Implement new `MomentumQNGOptimizer` compatible with (q)jit (#7606)
The goal here is to implement a new Jax-based version of the
`MomentumQNGOptimizer` compatible with `jax.jit` and `qml.qjit`. This PR
depends on the q(jit) compatible version of the standard `QNGOptimizer`
(see #7452 for more context).

Note that this optimization algorithm requires to keep track of a
non-trivial internal state: in the current Autograd-based
implementation, this state is stored in the optimizer's `accumulation`
attribute as a `pennylane.numpy` array (see
[here](https://github.com/PennyLaneAI/pennylane/blob/master/pennylane/optimize/momentum_qng.py#L122)).
However, in this Jax-based version, the state needs to be carried over
at each optimization step as it happens for the native `optax`
optimizers. This is required to avoid having a Jax tracer as a class
attribute when using (q)jit compilation.

[sc-92558]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>