Tach for `measurements` module (#7945)
**Context:**

We are trying to untangle the circular dependencies in pennylane.

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

Classical shadows have a dependence on the tertiary `shadows` module.
I'm just silencing that one. We should consider moving the measurements
to the `shadows` module instead.

`MeasurementProcess.expand` also has a dependence on `QuantumScript`.
That will be deprecated soon though.

**Related GitHub Issues:**

[sc-96131]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>