Update value of `rng_salt` to 0.42.0 (#7727)
------------------------------------------------------------------------------------------------------------

**Context:**
Set this ready for review after 0.42 released

**Description of the Change:**
See
https://github.com/PennyLaneAI/pennylane/pull/7727#issuecomment-3084849955
for a detailed list of the 9 fails this time.
TL;DR
Some are shot vector tests which should remove accuracy validations as
last time; one xfailed now passed, and one fluke; others are repeating
local-salted ones, which requrie separate PR's to be revised and fixed.
For now, we will change their salts.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**[sc-93829]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>