[Bug Fix Release] raise warning if `jax>0.6.2` is installed (#7949)
**Context:**

Jax released a [new
version](https://github.com/jax-ml/jax/releases/tag/jax-v0.7.0) with a
lot of breaking changes. Unfortunately our `v0.42` release is not
compatible with this version and so there are runtime errors.

ℹ️ : This PR was created by branching the v0.42.0 tag.

**Description of the Change:**

- The fix will check if the user has `jax` and if they do it will
perform a version check to ensure compatibility.

```python
>>> import pennylane as qml
```
```
~/Documents/pennylane/pennylane/__init__.py:29: RuntimeWarning: PennyLane is not yet compatible with JAX versions > 0.6.2. You have version 0.7.0 installed. Please downgrade JAX to <=0.6.2 to avoid runtime errors.
```

- We need to install the stable version of `pennylane-lightning` and
`pennylane-catalyst` instead of the latest, because the latest version
of these packages is now incompatible with this branch (they are using
the `pennylane.exceptions` module for some custom exceptions). This
change will not be merged into master.

**Benefits:** Better UI.

**Possible Drawbacks:** None.

[sc-96104]

---------

Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>
Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>