Use opaque syntax for the measurement-plane attribute of the mbqc dialect (#7957)
**Context:** The generic format of attributes in xDSL and in MLIR need
to agree in order for the output of a Python compiler pass to be used as
input to Catalyst. As such, the measurement-plane attribute of the
`mbqc` dialect needs to be output in the "opaque" format. For example,
the generic format of an `mbqc.measure_in_basis` op should be

```mlir
%res:2 = "mbqc.measure_in_basis"(%qbit, %angle) <{plane = #mbqc<measurement_plane XY>, postselect = 0 : i32}> : (!quantum.bit, f64) -> (i1, !quantum.bit)
```

Previously, without the opaque format, the `plane` attribute was output
as

```mlir
<{plane = #mbqc.measurement_planeXY}>
```

which is incompatible with the attribute as defined in Catalyst.

**Description of the Change:** The class `MeasurementPlaneAttr` now also
inherits from the `SpacedOpaqueSyntaxAttribute` class so that it uses
the opaque format for its generic output.

**Benefits:** Unblocks usage of `mbqc.measure_in_basis` ops in the
python compiler and in Catalyst.