Deprecate `MeasurementProcess.expand` (#7953)
**Context:**

`MeasurementProcess.expand` gives measurements a dependency on
`QuantumScript`, even though it should be the other way around. This
leads to a circular dependency.

Also, the method is only use in `expand_tape`, which has also fallen out
of favour.

**Description of the Change:**

Deprecates `MeasurementProcess.expand()`

**Benefits:**

Cleaner abstractions.

**Possible Drawbacks:**

It's still a deprecation, even if a harmless one.

**Related GitHub Issues:**

[sc-95802]

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>