fixing 1 qubit-size SemiAdder (#7940)
**Context:**

the SemiAdder was not working in the edge case of x + y defined with a
single qubit. This PR covers this situation

**Description of the Change:**

In this scenario, the semiAdder is a CNOT.
Note that in this case work_wires are not needed. So I had to change the
logic that was assuming we always have work wires

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>